+++ MP3/profiler.h
@@ -4,27 +4,31 @@
 #include <linux/kernel.h>
 #include <linux/list.h>
 #include <linux/module.h>
+#include <linux/workqueue.h>
+#include <linux/vmalloc.h>
+#include <linux/slab.h>
+#include <linux/workqueue.h>
 
 struct mp3_task_struct{
     struct task_struct* linux_task;
     struct list_head list;
     int pid;
-    uint32_t last_jiffies;
-    uint32_t utime;
-    uint32_t stime;
-    uint32_t maj_flt;
-    uint32_t min_flt;
+    unsigned long last_jiffies;
+    unsigned long utime;
+    unsigned long stime;
+    unsigned long maj_flt;
+    unsigned long min_flt;
 };
 
 extern struct mp3_task_struct tl;
 extern spinlock_t lock;
-extern long *vmem_buf;
+extern unsigned long *vmem_buf;
 
 void wq_handler(struct work_struct *w);
 
 void TaskList_init(void);
 void TaskList_destroy(void);
-int Task_register(uint32_t pid);
-int Task_deregister(uint32_t pid);
+int Task_register(int pid);
+int Task_deregister(int pid);
 
 #endif
